pyproject.toml
[project]
name = "rag-practice-01"
version = "0.1.0"
description = "RAG (Retrieval-Augmented Generation) practice project with various ML/NLP libraries"
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    # 언어 및 자연어 처리 관련 패키지
    "langchain>=0.1.0,<1.0.0",
    "langchain-core>=0.1.0,<1.0.0",
    "langchain-experimental>=0.1.0,<1.0.0",
    "langchain-community>=0.1.0,<1.0.0",
    "langchain-openai>=0.1.0,<1.0.0",
    "langchain-anthropic>=0.1.0,<1.0.0",
    "langchain-text-splitters>=0.1.0,<1.0.0",
    "langchain-elasticsearch>=0.1.0,<1.0.0",
    "langchain-chroma>=0.1.0,<1.0.0",
    "langchain-cohere>=0.1.0,<1.0.0",
    "langchain-milvus>=0.1.0,<1.0.0",
    "langchain-google-genai>=2.0.0,<3.0.0",
    "langchain-huggingface>=0.1.0,<1.0.0",
    "langchain-azure-ai>=0.1.0,<1.0.0",
    "langchain-teddynote>=0.1.0,<1.0.0",
    "langchainhub>=0.1.0,<1.0.0",
    "langgraph>=0.1.0,<1.0.0",
    "langsmith>=0.1.0,<1.0.0",
    "huggingface-hub>=0.20.0,<1.0.0",
    "openai>=1.0.0,<2.0.0",
    "deepl>=1.0.0,<2.0.0",
    "kiwipiepy>=0.15.0,<1.0.0",
    "konlpy>=0.6.0,<1.0.0",
    # 데이터 처리 및 분석 관련 패키지
    "pandas>=2.0.0,<3.0.0",
    "rank-bm25>=0.2.0,<1.0.0",
    # 데이터베이스 및 캐시 관련 패키지
    "redis>=5.0.0,<6.0.0",
    "chromadb>=0.4.0,<1.0.0",
    # PDF 및 파일 처리 관련 패키지
    "pymupdf>=1.23.0,<2.0.0",
    "pypdf>=4.0.0,<5.0.0",
    "pdfplumber>=0.10.0,<1.0.0",
    "pdfminer.six==20231228",
    "pymupdf4llm>=0.0.5,<1.0.0",
    # 시각화 및 사용자 인터페이스 관련 패키지
    "matplotlib>=3.7.0,<4.0.0",
    "streamlit>=1.28.0,<2.0.0",
    "jupyter>=1.0.0,<2.0.0",
    "notebook>=7.0.0,<8.0.0",
    # 딥러닝 및 머신러닝 관련 패키지
    "torch==2.2.0",
    "torchvision>=0.17.0,<1.0.0",
    "faiss-cpu>=1.7.0,<2.0.0",
    "open-clip-torch>=2.20.0,<3.0.0",
    # 기타 유틸리티 및 필수 패키지
    "python-dotenv>=1.0.0,<2.0.0",
    "pydantic>=2.5.0,<3.0.0",
    "lxml>=5.0.0,<6.0.0",
    "pillow>=10.0.0,<11.0.0",
    "lark>=1.1.0,<2.0.0",
    "ragas>=0.1.0,<1.0.0",
    "unstructured[all-docs]>=0.11.0,<1.0.0",
    "arxiv>=2.0.0,<3.0.0",
    "tiktoken>=0.5.0,<1.0.0",
    "tenacity>=8.2.0,<9.0.0",
    "pymilvus>=2.3.0,<3.0.0",
    "google-search-results>=2.4.0,<3.0.0",
    "protobuf>=3.20.0,<5.0.0",
    "sqlalchemy>=2.0.0,<3.0.0",
    "llama-index-core>=0.10.0,<1.0.0",
    "llama-parse>=0.4.0,<1.0.0",
    "llama-index-readers-file>=0.1.0,<1.0.0",
    "flashrank>=0.2.0,<1.0.0",
    "docx2txt>=0.8.0,<1.0.0",
    "nest-asyncio>=1.5.0,<2.0.0",
    "rapidocr-onnxruntime>=1.3.0,<2.0.0",
    "seaborn>=0.12.0,<1.0.0",
    "grandalf>=0.8.0,<1.0.0",
    "rouge-score>=0.1.0,<1.0.0",
    "langchain-ollama>=0.1.0,<1.0.0",
    "mypy>=1.7.0,<2.0.0",
    "pinecone>=5.0.0,<6.0.0",
    "wikipedia>=1.4.0,<2.0.0",
    "scikit-learn>=1.3.0,<2.0.0",
    "psycopg2-binary>=2.9.10",
    "langchain-upstage>=0.7.3",
    "pgvector>=0.4.1",
]

[dependency-groups]
dev = [
    "mypy>=1.18.1",
    "pylsp-mypy>=0.7.0",
    "python-lsp-server>=1.13.1",
    "sqlalchemy-stubs>=0.4",
    "types-sqlalchemy>=1.4.53.38",
]
